+++ src/Data/Morpheus/Execution/Server/Resolve.hs
@@ -38,7 +38,7 @@ import           Data.Morpheus.Schema.SchemaAPI                      (defaultTyp
 import           Data.Morpheus.Types.GQLType                         (GQLType (CUSTOM))
 import           Data.Morpheus.Types.Internal.AST.Operation          (Operation (..), ValidOperation)
 import           Data.Morpheus.Types.Internal.Data                   (DataFingerprint (..), DataTyCon (..),
-                                                                      DataTypeLib (..), MUTATION, OperationKind (..),
+                                                                      DataTypeLib (..), MUTATION, OperationType (..),
                                                                       QUERY, SUBSCRIPTION, initTypeLib)
 import           Data.Morpheus.Types.Internal.Resolver               (GQLRootResolver (..), Resolver (..), ResponseT,
                                                                       toResponseRes)
@@ -103,10 +103,10 @@ streamResolver root@GQLRootResolver {queryResolver, mutationResolver, subscripti
         query <- parseGQL request >>= validateRequest schema FULL_VALIDATION
         Right (schema, query)
     ----------------------------------------------------------
-    execOperator (schema, operation@Operation {operationKind = Query}) =
+    execOperator (schema, operation@Operation {operationType = Query}) =
         toResponseRes (encodeQuery (schemaAPI schema) queryResolver operation)
-    execOperator (_, operation@Operation {operationKind = Mutation}) = toResponseRes (encodeMutation mutationResolver operation)
-    execOperator (_, operation@Operation {operationKind = Subscription}) = response
+    execOperator (_, operation@Operation {operationType = Mutation}) = toResponseRes (encodeMutation mutationResolver operation)
+    execOperator (_, operation@Operation {operationType = Subscription}) = response
         where
          response = toResponseRes (encodeSubscription subscriptionResolver operation)